requirements.txt
fastapi
uvicorn
motor
python-dotenv
dotenv
bs4
pdfminer.six
bcrypt
